ipf: Resume fragments in same action list.
Once fragments are reassembled and go through conntrack,
the fragments need to resume processing in the same action
list.  There is a semantic requirement that the list have a
single conntrack action so this specifies the resume point in
the list as well. A memcmp is used to compare lists as padding
is zeroed out, therefore having predictable values.

Signed-off-by: Darrell Ball <dlu998@gmail.com>